Refactored code for TestSnapshot 1.x & 2.x. Also refactored code for TestAsyncSnpshot by creating AbstractTestAsycSnapshot.
