+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x-integration-tests/src/test/java/com/google/cloud/bigtable/hbase/async/TestAsyncSnapshots.java
@@ -16,14 +16,8 @@
 
 package com.google.cloud.bigtable.hbase.async;
 
-import static com.google.cloud.bigtable.hbase.test_env.SharedTestEnvRule.COLUMN_FAMILY;
-
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.regex.Pattern;
 
@@ -31,101 +25,56 @@ import org.apache.hadoop.hbase.HColumnDescriptor;
 import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.AsyncAdmin;
-import org.apache.hadoop.hbase.client.Put;
-import org.apache.hadoop.hbase.client.Result;
-import org.apache.hadoop.hbase.client.ResultScanner;
-import org.apache.hadoop.hbase.client.Scan;
-import org.apache.hadoop.hbase.client.SnapshotDescription;
-import org.apache.hadoop.hbase.client.Table;
-import org.apache.hadoop.hbase.util.Bytes;
+import org.apache.hadoop.hbase.client.ColumnFamilyDescriptorBuilder;
+import org.apache.hadoop.hbase.client.TableDescriptor;
+import org.apache.hadoop.hbase.client.TableDescriptorBuilder;
 import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-import com.google.cloud.bigtable.hbase.Logger;
+import com.google.cloud.bigtable.hbase.AbstractTestSnapshot;
 import com.google.cloud.bigtable.hbase.test_env.SharedTestEnvRule;
 
 @RunWith(JUnit4.class)
-public class TestAsyncSnapshots  extends AbstractAsyncTest {
-	protected Logger logger = new Logger(this.getClass());
-	final byte[] QUALIFIER = dataHelper.randomData("TestAsyncSnapshots");
+public class TestAsyncSnapshots extends AbstractTestSnapshot {
 
-  private final TableName tableName = sharedTestEnv.newTestTableName();
-  private final TableName anotherTableName = sharedTestEnv.newTestTableName();
-  // The maximum size of a table id or snapshot id is 50. newTestTableName().size() can approach 50.
-  // Make sure that the snapshot name and cloned table are within the 50 character limit
-  private final String snapshotName = tableName.getNameAsString().substring(40) + "_snp";
-  private final String anotherSnapshotName = 
-      anotherTableName.getNameAsString().substring(40) + "_snp";
-  private final TableName clonedTableName =
-      TableName.valueOf(tableName.getNameAsString().substring(40) + "_clone");
+  @Before
+  public void setup() throws Exception {
+    descriptor = new HTableDescriptor(anotherTableName);
+    descriptor.addFamily(new HColumnDescriptor(SharedTestEnvRule.COLUMN_FAMILY));
+    createTable(anotherTableName);
+  }
 
   @After
-  public void cleanup() {
-    if (sharedTestEnv.isBigtable() && !enableTestForBigtable()) {
-      return;
-    }
-    try{
-      AsyncAdmin asyncAdmin = getAsyncAdmin();
-      delete(asyncAdmin, tableName);
-      delete(asyncAdmin, anotherTableName);
-      delete(asyncAdmin, clonedTableName);
-      asyncAdmin.listSnapshots().thenApply(r->{
-        for (SnapshotDescription snapshotDescription : r) {
-        	asyncAdmin.deleteSnapshot(snapshotDescription.getName());
-        }
-        return null;
-      });
-    } catch (Exception e) {
-      e.printStackTrace();
-    }
-  }
-  
-  protected boolean enableTestForBigtable() {
-    return false;
-  }
-  
-  private void delete(AsyncAdmin asyncAdmin, TableName tableName) throws IOException {
-    if(asyncAdmin.tableExists(tableName) != null) {
-    	asyncAdmin.disableTable(tableName);
-    	asyncAdmin.deleteTable(tableName);
-    }
+  public void tearDown() throws IOException, InterruptedException, ExecutionException {
+    getAsyncAdmin().disableTable(anotherTableName).get();
+    getAsyncAdmin().deleteTable(anotherTableName).get();
   }
   
   private AsyncAdmin getAsyncAdmin() throws InterruptedException, ExecutionException {
     return AbstractAsyncTest.getAsyncConnection().getAdmin();
   }
   
-  @Test
-  public void testDeleteSnapshots() throws InterruptedException, ExecutionException {
-  	AsyncAdmin asyncAdmin = getAsyncAdmin();
-  	asyncAdmin.deleteSnapshots();
-  	
-  	//To Test delete snapshot with Pattern
-  	asyncAdmin.snapshot(anotherSnapshotName, anotherTableName);
-  	final Pattern allSnapshots = Pattern.compile(anotherSnapshotName + ".*");
-  	asyncAdmin.deleteSnapshots(allSnapshots);
-  }
-  
-  @Test
-  public void testDeleteTableSnapshots() throws InterruptedException, ExecutionException {
-  	AsyncAdmin asyncAdmin = getAsyncAdmin();
-  	//To Test delete snapshot with Pattern
-  	asyncAdmin.snapshot(anotherSnapshotName, anotherTableName);
-  	final Pattern allSnapshots = Pattern.compile(anotherSnapshotName + ".*");
-  	asyncAdmin.deleteTableSnapshots(allSnapshots);
-  }
-  
-  @Test
-  public void testcloneSnapshot() throws InterruptedException, ExecutionException {    
-  	AsyncAdmin asyncAdmin = getAsyncAdmin();
-  	asyncAdmin.cloneSnapshot(snapshotName, clonedTableName);;
+  @Override
+  protected void createTable(TableName tableName) throws IOException {
+    try{
+      getAsyncAdmin().createTable(createDescriptor(tableName)).get();
+    } catch (InterruptedException | ExecutionException e) {
+      e.printStackTrace();
+    }
   }  
   
+  private TableDescriptor createDescriptor(TableName tableName) {
+    return TableDescriptorBuilder.newBuilder(tableName)
+        .addColumnFamily(ColumnFamilyDescriptorBuilder.newBuilder(SharedTestEnvRule.COLUMN_FAMILY).build())
+        .build();
+  }
+
   @Test
-  public void testSnapshot() throws IOException {
+  public void testAsyncSnapshot() throws IOException {
     if (sharedTestEnv.isBigtable() && !Boolean.getBoolean("perform.snapshot.test")) {
       return;
     }
@@ -135,7 +84,7 @@ public class TestAsyncSnapshots  extends AbstractAsyncTest {
     	asyncAdmin.createTable(
         new HTableDescriptor(tableName).addFamily(new HColumnDescriptor(SharedTestEnvRule.COLUMN_FAMILY)));
 
-      Map<String, Long> values = createAndPopulateTable();
+      Map<String, Long> values = createAndPopulateTable(tableName);
       checkSnapshotCount(asyncAdmin, 0);
       asyncAdmin.snapshot(snapshotName, tableName);
       checkSnapshotCount(asyncAdmin, 1);
@@ -151,20 +100,6 @@ public class TestAsyncSnapshots  extends AbstractAsyncTest {
         e.printStackTrace();
       }
   }
-
-  protected void validateClone(Map<String, Long> values) throws IOException {
-	    try (Table clonedTable = getConnection().getTable(clonedTableName);
-	        ResultScanner scanner = clonedTable.getScanner(new Scan())){
-	      for (Result result : scanner) {
-	        String row = Bytes.toString(result.getRow());
-	        Long expected = values.get(row);
-	        Long found = Bytes.toLong(result.getValue(COLUMN_FAMILY, QUALIFIER));
-	        Assert.assertEquals("row " + row + " not equal", expected, found);
-	        values.remove(row);
-	      }
-	    }
-	    Assert.assertTrue("There were missing keys.", values.isEmpty());
-  }
   
 	private void checkSnapshotCount(AsyncAdmin asyncAdmin, int count) {
 			asyncAdmin.listSnapshots().thenApply(r->{
@@ -173,19 +108,4 @@ public class TestAsyncSnapshots  extends AbstractAsyncTest {
     	return null;
 		});
 	}
-  private Map<String, Long> createAndPopulateTable() throws IOException {
-	    Map<String, Long> values = new HashMap<>();
-	    try (Table table = getConnection().getTable(tableName)) {
-	      values.clear();
-	      List<Put> puts = new ArrayList<>();
-	      for (long i = 0; i < 10; i++) {
-	        final UUID rowKey = UUID.randomUUID();
-	        byte[] row = Bytes.toBytes(rowKey.toString());
-	        values.put(rowKey.toString(), i);
-	        puts.add(new Put(row).addColumn(SharedTestEnvRule.COLUMN_FAMILY, QUALIFIER, Bytes.toBytes(i)));
-	      }
-	      table.put(puts);
-	    }
-	    return values;
-  }
 }
